+++ src/rpcblockchain.cpp
@@ -3,9 +3,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-
-
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
 #include "main.h"
 #include "sync.h"
