+++ src/rpcnet.cpp
@@ -2,9 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-
-
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
 #include "net.h"
 #include "netbase.h"
 #include "protocol.h"
