+++ src/test/rpc_tests.cpp
@@ -1,4 +1,5 @@
-#include "bitcoinrpc.h"
+#include "rpcserver.h"
+#include "rpcclient.h"
 
 #include "base58.h"
